[Refactor] 인기글관련 부분 클래스 분리 #38
