+++ src/main/java/com/example/starhub/controller/MeetingController.java
@@ -132,50 +132,4 @@ public class MeetingController implements MeetingControllerDocs {
                 .status(ResponseCode.SUCCESS_GET_CONFIRMED_MEMBERS.getStatus().value())
                 .body(new ResponseDto<>(ResponseCode.SUCCESS_GET_CONFIRMED_MEMBERS, res));
     }
-
-    /**
-     * 인기글 페이지 - 프로젝트 인기글 3개
-     *
-     */
-    @GetMapping("/popular/projects")
-    public ResponseEntity<ResponseDto<List<MeetingSummaryResponseDto>>> getPopularProjects(
-            @Parameter(hidden = true) @AuthenticationPrincipal CustomUserDetails customUserDetails) {
-
-        // 익명 사용자일 경우 null 전달, 인증된 사용자일 경우 customUserDetails 전달
-        String username = customUserDetails != null ? customUserDetails.getUsername() : null;
-
-        List<MeetingSummaryResponseDto> res = meetingService.getPopularProjects(username);
-        return ResponseEntity
-                .status(ResponseCode.SUCCESS_GET_POPULAR_PROJECTS.getStatus().value())
-                .body(new ResponseDto<>(ResponseCode.SUCCESS_GET_POPULAR_PROJECTS, res));
-    }
-
-    /**
-     * 인기글 페이지 - 스터디 인기글 3개
-     */
-    @GetMapping("/popular/studies")
-    public ResponseEntity<ResponseDto<List<MeetingSummaryResponseDto>>> getPopularStudies(
-            @Parameter(hidden = true) @AuthenticationPrincipal CustomUserDetails customUserDetails) {
-
-        String username = customUserDetails != null ? customUserDetails.getUsername() : null;
-        List<MeetingSummaryResponseDto> res = meetingService.getPopularStudies(username);
-        return ResponseEntity
-                .status(ResponseCode.SUCCESS_GET_POPULAR_STUDIES.getStatus().value())
-                .body(new ResponseDto<>(ResponseCode.SUCCESS_GET_POPULAR_STUDIES, res));
-    }
-
-    /**
-     * 인기글 페이지 - 마감입박 인기글 3개
-     */
-    @GetMapping("/popular/expiring")
-    public ResponseEntity<ResponseDto> getExpiringPopularMeetings(
-            @Parameter(hidden = true) @AuthenticationPrincipal CustomUserDetails customUserDetails) {
-
-        String username = customUserDetails != null ? customUserDetails.getUsername() : null;
-        List<MeetingSummaryResponseDto> res = meetingService.getExpiringPopularMeetings(username);
-        return ResponseEntity
-                .status(ResponseCode.SUCCESS_GET_POPULAR_EXPIRING.getStatus().value())
-                .body(new ResponseDto<>(ResponseCode.SUCCESS_GET_POPULAR_EXPIRING, res));
-    }
-
 }
